requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
beautifulsoup4==4.12.2
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
